backend/requirements.txt
fastapi
uvicorn
pydantic
faker
pandas
redis
rq
paho-mqtt
kafka-python
pika
requests
toml
dict2xml